+++ fleetpings/constants.py
@@ -2,6 +2,4 @@
 Constants
 """
 
-DEFAULT_EMBED_COLOR = "#FAA61A"
-
 DISCORD_WEBHOOK_REGEX = r"https:\/\/discord\.com\/api\/webhooks\/[\d]+\/[a-zA-Z0-9_-]+$"