+++ fleetpings/migrations/0004_auto_20200915_1617.py
@@ -5,7 +5,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("fleetpings", "0003_fleetdoctrine_link"),
     ]