+++ fleetpings/migrations/0009_remove_webhook_is_embedded_remove_webhook_type_and_more.py
@@ -5,7 +5,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("fleetpings", "0008_auto_20210114_1733"),
     ]