+++ fleetpings/migrations/0012_migrate_app_settings_from_local_py_to_database.py
@@ -1,0 +1,54 @@
+# Django
+from django.db import migrations
+
+# AA Fleet Pings
+from fleetpings.utils import clean_setting
+
+use_doctrines_from_fittings_module = clean_setting(
+    "AA_FLEETPINGS_USE_DOCTRINES_FROM_FITTINGS_MODULE", False
+)
+
+webhook_verification = clean_setting("AA_FLEETPINGS_WEBHOOK_VERIFICATION", True)
+
+
+def on_migrate(apps, schema_editor):
+    """
+    Create default settings on migration
+    :param apps:
+    :param schema_editor:
+    :return:
+    """
+
+    Setting = apps.get_model("fleetpings", "Setting")
+    db_alias = schema_editor.connection.alias
+
+    Setting.objects.using(db_alias).create(
+        pk=1,
+        use_doctrines_from_fittings_module=use_doctrines_from_fittings_module,
+        webhook_verification=webhook_verification,
+    )
+
+
+def on_migrate_zero(apps, schema_editor):
+    """
+    Remove default settings on migratio to zero
+    :param apps:
+    :param schema_editor:
+    :return:
+    """
+
+    Setting = apps.get_model("fleetpings", "Setting")
+    db_alias = schema_editor.connection.alias
+    Setting.objects.using(db_alias).delete()
+
+
+class Migration(migrations.Migration):
+    """
+    Run migrations
+    """
+
+    dependencies = [
+        ("fleetpings", "0011_settings_and_verbose_names"),
+    ]
+
+    operations = [migrations.RunPython(on_migrate, on_migrate_zero)]