+++ fleetpings/models.py
@@ -16,14 +16,11 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 # AA Fleet Pings
-from fleetpings.app_settings import (
-    AA_FLEETPINGS_WEBHOOK_VERIFICATION,
-    discord_service_installed,
-)
+from fleetpings.app_settings import discord_service_installed
+from fleetpings.constants import DISCORD_WEBHOOK_REGEX
+from fleetpings.managers import SettingManager
 
 # Check if the Discord service is active
-from fleetpings.constants import DISCORD_WEBHOOK_REGEX
-
 if discord_service_installed():
     # Alliance Auth
     from allianceauth.services.modules.discord.models import DiscordUser
@@ -40,7 +37,7 @@ def _get_discord_group_info(ping_target: Group) -> dict:
 
     if not discord_service_installed():
         raise ValidationError(
-            _("You might want to install the Discord service first ...")
+            _("You might want to install the Discord service first …")
         )
 
     try:
@@ -49,11 +46,47 @@ def _get_discord_group_info(ping_target: Group) -> dict:
         raise ValidationError(
             _("Are you sure you have your Discord linked to your Alliance Auth?")
         ) from http_error
-    else:
-        if not discord_group_info:
-            raise ValidationError(_("This group has not been synced to Discord yet."))
 
-        return discord_group_info
+    if not discord_group_info:
+        raise ValidationError(_("This group has not been synced to Discord yet."))
+
+    return discord_group_info
+
+
+class SingletonModel(models.Model):
+    """
+    SingletonModel
+    """
+
+    class Meta:  # pylint: disable=too-few-public-methods
+        """
+        Model meta definitions
+        """
+
+        abstract = True
+
+    def save(self, *args, **kwargs):
+        """
+        Save action
+        :param args:
+        :param kwargs:
+        :return:
+        """
+
+        if self.__class__.objects.count():
+            self.pk = self.__class__.objects.first().pk
+
+        super().save(*args, **kwargs)
+
+    def delete(self, *args, **kwargs):
+        """
+        Delete action
+        :param args:
+        :param kwargs:
+        :return:
+        """
+
+        pass  # pylint: disable=unnecessary-pass
 
 
 class AaFleetpings(models.Model):
@@ -78,19 +111,24 @@ class FleetComm(models.Model):
     """
 
     name = models.CharField(
-        max_length=255, unique=True, help_text=_("Short name to identify this comms")
+        max_length=255,
+        unique=True,
+        help_text=_("Short name to identify this comms"),
+        verbose_name=_("Name"),
     )
 
     notes = models.TextField(
         null=True,
         blank=True,
         help_text=_("You can add notes about this configuration here if you want"),
+        verbose_name=_("Notes"),
     )
 
     is_enabled = models.BooleanField(
         default=True,
         db_index=True,
         help_text=_("Whether this comms is enabled or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def __str__(self) -> str:
@@ -114,14 +152,18 @@ class FleetDoctrine(models.Model):
 
     # Doctrine name
     name = models.CharField(
-        max_length=255, unique=True, help_text=_("Short name to identify this doctrine")
+        max_length=255,
+        unique=True,
+        help_text=_("Short name to identify this doctrine"),
+        verbose_name=_("Name"),
     )
 
     # Link to your doctrine
     link = models.CharField(
         max_length=255,
-        help_text=_("A link to a doctrine page for this doctrine if you have."),
         blank=True,
+        help_text=_("A link to a doctrine page for this doctrine if you have."),
+        verbose_name=_("Doctrine link"),
     )
 
     # Restrictions
@@ -129,7 +171,8 @@ class FleetDoctrine(models.Model):
         Group,
         blank=True,
         related_name="fleetdoctrine_require_groups",
-        help_text=_("Restrict this doctrine to the following group(s) ..."),
+        help_text=_("Restrict this doctrine to the following group(s) …"),
+        verbose_name=_("Group restrictions"),
     )
 
     # Doctrine notes
@@ -137,6 +180,7 @@ class FleetDoctrine(models.Model):
         null=True,
         blank=True,
         help_text=_("You can add notes about this configuration here if you want"),
+        verbose_name=_("Notes"),
     )
 
     # Is doctrine active
@@ -144,6 +188,7 @@ class FleetDoctrine(models.Model):
         default=True,
         db_index=True,
         help_text=_("Whether this doctrine is enabled or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def clean(self):
@@ -189,6 +234,7 @@ class FormupLocation(models.Model):
         max_length=255,
         unique=True,
         help_text=_("Short name to identify this formup location"),
+        verbose_name=_("Name"),
     )
 
     # Formup location notes
@@ -196,13 +242,15 @@ class FormupLocation(models.Model):
         null=True,
         blank=True,
         help_text=_("You can add notes about this configuration here if you want"),
+        verbose_name=_("Notes"),
     )
 
-    # Is formup location active
+    # Is formup location active?
     is_enabled = models.BooleanField(
         default=True,
         db_index=True,
         help_text=_("Whether this formup location is enabled or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def __str__(self) -> str:
@@ -235,6 +283,7 @@ class DiscordPingTargets(models.Model):
                 "(Note: This must be an Auth group that is synced to Discord.)"
             )
         ),
+        verbose_name=_("Group name"),
     )
 
     # Discord group id
@@ -243,6 +292,7 @@ class DiscordPingTargets(models.Model):
         unique=True,
         blank=True,
         help_text=_("ID of the Discord role to ping"),
+        verbose_name=_("Discord ID"),
     )
 
     # Restrictions
@@ -250,7 +300,8 @@ class DiscordPingTargets(models.Model):
         Group,
         blank=True,
         related_name="discord_role_require_groups",
-        help_text=_("Restrict ping rights to the following group(s) ..."),
+        help_text=_("Restrict ping rights to the following group(s) …"),
+        verbose_name=_("Group restrictions"),
     )
 
     # Notes
@@ -258,13 +309,15 @@ class DiscordPingTargets(models.Model):
         null=True,
         blank=True,
         help_text=_("You can add notes about this configuration here if you want"),
+        verbose_name=_("Notes"),
     )
 
-    # Is this group active
+    # Is this group active?
     is_enabled = models.BooleanField(
         default=True,
         db_index=True,
         help_text=_("Whether this formup location is enabled or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def clean(self):
@@ -315,13 +368,15 @@ class FleetType(models.Model):
         max_length=255,
         unique=True,
         help_text=_("Short name to identify this fleet type"),
+        verbose_name=_("Name"),
     )
 
     # Embed color
     embed_color = models.CharField(
         max_length=7,
         blank=True,
-        help_text=_("Hightlight color for the embed"),
+        help_text=_("Highlight color for the embed"),
+        verbose_name=_("Embed color"),
     )
 
     # Restrictions
@@ -329,7 +384,8 @@ class FleetType(models.Model):
         Group,
         blank=True,
         related_name="fleettype_require_groups",
-        help_text=_("Restrict this fleet type to the following group(s) ..."),
+        help_text=_("Restrict this fleet type to the following group(s) …"),
+        verbose_name=_("Group restrictions"),
     )
 
     # Fleet type notes
@@ -337,6 +393,7 @@ class FleetType(models.Model):
         null=True,
         blank=True,
         help_text=_("You can add notes about this configuration here if you want"),
+        verbose_name=_("Notes"),
     )
 
     # Is this fleet type enabled
@@ -344,6 +401,7 @@ class FleetType(models.Model):
         default=True,
         db_index=True,
         help_text=_("Whether this fleet type is enabled or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def __str__(self) -> str:
@@ -370,6 +428,7 @@ class Webhook(models.Model):
         max_length=255,
         unique=True,
         help_text=_("Name of the channel this webhook posts to"),
+        verbose_name=_("Discord channel"),
     )
 
     # Wehbook url
@@ -382,6 +441,7 @@ class Webhook(models.Model):
                 "https://discord.com/api/webhooks/123456/abcdef"
             )
         ),
+        verbose_name=_("Webhook URL"),
     )
 
     # Restrictions
@@ -389,7 +449,8 @@ class Webhook(models.Model):
         Group,
         blank=True,
         related_name="webhook_require_groups",
-        help_text=_("Restrict ping rights to the following group(s) ..."),
+        help_text=_("Restrict ping rights to the following group(s) …"),
+        verbose_name=_("Group restrictions"),
     )
 
     # Webhook notes
@@ -397,13 +458,15 @@ class Webhook(models.Model):
         null=True,
         blank=True,
         help_text=_("You can add notes about this webhook here if you want"),
+        verbose_name=_("Notes"),
     )
 
-    # Is it enabled
+    # Is it enabled?
     is_enabled = models.BooleanField(
         default=True,
         db_index=True,
         help_text=_("Whether this webhook is active or not"),
+        verbose_name=_("Is enabled"),
     )
 
     def __str__(self) -> str:
@@ -424,11 +487,10 @@ class Webhook(models.Model):
         :return:
         """
 
-        # Check if it's an actual webhook url if the verification setting is set.
-        if (
-            not re.match(DISCORD_WEBHOOK_REGEX, self.url)
-            and AA_FLEETPINGS_WEBHOOK_VERIFICATION
-        ):
+        # Check if it's an actual Discord Webhook URL if the verification setting is set.
+        if not re.match(
+            DISCORD_WEBHOOK_REGEX, self.url
+        ) and Setting.objects.get_setting(Setting.Field.WEBHOOK_VERIFICATION):
             raise ValidationError(
                 _(
                     "Invalid webhook URL. The webhook URL you entered does not match "
@@ -438,3 +500,92 @@ class Webhook(models.Model):
             )
 
         super().clean()
+
+
+class Setting(SingletonModel):
+    """
+    Default forum settings
+    """
+
+    class Field(models.TextChoices):
+        """
+        Choices for Setting.Field
+        """
+
+        USE_DEFAULT_FLEET_TYPES = "use_default_fleet_types", _(
+            "Use default fleet types"
+        )
+        USE_DEFAULT_PING_TARGETS = "use_default_ping_targets", _(
+            "Use default ping targets"
+        )
+        USE_DOCTRINES_FROM_FITTINGS_MODULE = "use_doctrines_from_fittings_module", _(
+            "Use Doctrines from Fittings module"
+        )
+        WEBHOOK_VERIFICATION = "webhook_verification", _("Verify Webhooks")
+        DEFAULT_EMBED_COLOR = "default_embed_color", _("Default embed color")
+
+    use_default_fleet_types = models.BooleanField(
+        default=True,
+        db_index=True,
+        help_text=_(
+            "Whether to use default fleet types. If checked, the default fleet types "
+            "(Roaming, Home Defense, StratOP, and CTA) will be added to the Fleet Type "
+            "dropdown."
+        ),
+        verbose_name=Field.USE_DEFAULT_FLEET_TYPES.label,
+    )
+
+    use_default_ping_targets = models.BooleanField(
+        default=True,
+        db_index=True,
+        help_text=_(
+            "Whether to use default ping targets. If checked, the default ping targets "
+            "(@everyone and @here) will be added to the Ping Target dropdown."
+        ),
+        verbose_name=Field.USE_DEFAULT_PING_TARGETS.label,
+    )
+
+    use_doctrines_from_fittings_module = models.BooleanField(
+        default=False,
+        db_index=True,
+        help_text=_(
+            "Whether to use the doctrines from the Fittings modules in the doctrine "
+            "dropdown. Note: The fittings module needs to be installed for this."
+        ),
+        verbose_name=Field.USE_DOCTRINES_FROM_FITTINGS_MODULE.label,
+    )
+
+    webhook_verification = models.BooleanField(
+        default=True,
+        db_index=True,
+        help_text=_(
+            "Whether to verify Webhooks URLs or not. Note: When unchecked Webhook URLs "
+            "will not be verified, so the app can be used with non-Discord Webhooks "
+            "as well. When disabling webhook verification and using non-Discord "
+            "webhooks, it is up to you to make sure your webhook understands a payload "
+            "that is formatted for Discord webhooks."
+        ),
+        verbose_name=Field.WEBHOOK_VERIFICATION.label,
+    )
+
+    default_embed_color = models.CharField(
+        default="#FAA61A",
+        max_length=7,
+        blank=True,
+        help_text=_("Default highlight color for the webhook embed."),
+        verbose_name=Field.DEFAULT_EMBED_COLOR.label,
+    )
+
+    objects = SettingManager()
+
+    class Meta:  # pylint: disable=too-few-public-methods
+        """
+        Meta definitions
+        """
+
+        default_permissions = ()
+        verbose_name = _("setting")
+        verbose_name_plural = _("settings")
+
+    def __str__(self) -> str:
+        return str(_("Fleet Pings Settings"))