+++ CHANGELOG.md
@@ -9,6 +9,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [In Development] - Unreleased
 
 
+## [2.22.0] - 2023-05-20
+
+### Changed
+
+- Migrated settings from `local.py` to the database
+- German translation updated
+- Russian translation updated
+- Ukrainian translation updated (not complete yet, currently at 90% translated)
+
+### Update Notes:
+
+After this update and successful migration, you can remove this app's settings from
+your `local.py`. All settings are now handled through the Django Admin Backend under
+`/admin/fleetpings/setting/`
+
+
 ## [2.21.0] - 2023-04-25
 
 ### Added
@@ -92,7 +108,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Templates cleaned up and re-organised
+- Templates cleaned up and re-organized
 - Minimum Requirement:
   - Alliance Auth >= 2.14.0
 
@@ -500,7 +516,7 @@ settings in the admin backend, since we just nuked them from the old app.
 - Configuration for pre-defined fleet comms, formup locations, and doctrines. These
   can be configured in your `local.py` via `AA_DISCORDFORMATTER_FLEET_COMMS`,
   `AA_DISCORDFORMATTER_FLEET_DOCTRINES` and
-  `AA_DISCORDFORMATTER_FLEET_FORMUP_LOCATIONS`. See [README](https://github.com/ppfeufer/aa-discord-ping-formatter#fleet-comms-formup-location-and-doctrine) for syntax.
+  `AA_DISCORDFORMATTER_FLEET_FORMUP_LOCATIONS`.
 
 
 ## [1.0.0] - 2020-07-16
@@ -518,7 +534,7 @@ settings in the admin backend, since we just nuked them from the old app.
 ### Fixed
 
 - Our Australian time travelers and everyone else who lives in the future (UTC+x) are
-  now able to pre-ping fleets that are coming up in 2 hours Eve time, which might
+  now able to pre-ping fleets that are coming up in 2-hours Eve time, which might
   still be in their past local time, depending on how far in the future they live. (#19)
 
 
@@ -529,7 +545,7 @@ settings in the admin backend, since we just nuked them from the old app.
 - Option to embed automatic pings via webhook (#13)
 - Embedded pings via webhook are now color coded. Pre-defined fleet types are by
   default (Roam = green, Home Defense = yellow, StratOP = orange, CTA = red), and custom
-  fleet types can be defined via settings (see [README](https://github.com/ppfeufer/aa-discord-ping-formatter#embed-webhook-pings))
+  fleet types can be defined via settings.
 
 ### Changed
 
@@ -551,8 +567,7 @@ settings in the admin backend, since we just nuked them from the old app.
 
 ### Added
 
-- Webhook group restrictions. Webhooks can now be restricted to certain groups (see
-  [README](https://github.com/ppfeufer/aa-discord-ping-formatter#adding-ping-channels)),
+- Webhook group restrictions. Webhooks can now be restricted to certain groups,
   so not everyone who has access to this module can ping through all webhooks.
   Webhooks without restrictions are accessible for all with access to the module.
   (Thanks to Exiom for bringing this up)